fix(panel): address round-2 review issues in skill diff
- Issue #1: replace blocking Command::output() with tokio::process::Command;
  read stdout up to MAX_DIFF_BYTES+1 via AsyncReadExt::take() so the cap
  is enforced while streaming, killing the child if the limit is exceeded;
  stderr is drained concurrently to prevent git blocking on a full pipe
- Issue #2: tighten the +/- line filter from !starts_with("+++") to
  !starts_with("+++ ") so content lines whose diff marker is +++/---
  (source line starting with ++) are captured, not silently dropped
- Issue #3: add parse_diff_git_path() to handle both unquoted and
  git-quoted (Unicode/special-char) diff --git header forms; quoted paths
  use a scan that respects backslash escapes
- Split skill-diff helpers + handler into src/panel/skill_diff.rs to keep
  handlers.rs within the 800-line limit (was ~787, would have grown to ~830)
- Add regression tests for the ++ content line and quoted-path cases
