+++ Cargo.toml
@@ -15,7 +15,7 @@ clap = { version = "4.5.37", features = ["derive"] }
 serde = { version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.140"
 thiserror = "2.0.12"
-tokio = { version = "1.45.1", features = ["macros", "rt-multi-thread", "signal"] }
+tokio = { version = "1.45.1", features = ["macros", "rt-multi-thread", "signal", "process", "io-util"] }
 ts-rs = { version = "12.0.1", features = ["chrono-impl"] }
 uuid = { version = "1.16.0", features = ["v4", "serde"] }
 walkdir = "2.5.0"
